rust: suggestions from clippy.

+++ src/libs/rust/base/src/tcu.rs
@@ -873,7 +873,7 @@ impl TCU {
         #[cfg(any(feature = "hw22", feature = "hw23"))]
         let regs = EXT_REGS + UNPRIV_REGS + (128 * EP_REGS) as usize;
         #[cfg(not(any(feature = "hw22", feature = "hw23")))]
-        let regs = EXT_REGS + UNPRIV_REGS as usize;
+        let regs = EXT_REGS + UNPRIV_REGS;
 
         let s = &s[0..cmp::min(s.len(), PRINT_REGS * mem::size_of::<Reg>() - 1)];
